+++ .github/workflows/pr__labeler.yml
@@ -102,7 +102,8 @@ jobs:
           number: ${{ github.event.issue.number }}
           header: 'slef-lgtm'
           message: |
-            Hi, are you the pull request owner?.
+            Hi, are you the pull request owner?
+            you can't lgtm this pull request.
 
       - name: Add Approved label
         if: |
