+++ package.json
@@ -18,10 +18,10 @@
     "test-cov": "jest --coverage",
     "styleguide": "styleguidist server",
     "styleguide:build": "styleguidist build",
-    "styleguide:proxy": "source bridge/environment.sh; styleguidist server & node bridge/index.js"
+    "styleguide:proxy": "source bridge/environment.sh; styleguidist server --config styleguide.config.proxy.js"
   },
   "dependencies": {
-    "@kubevirt-ui/kubevirt-api": "^0.0.30",
+    "@kubevirt-ui/kubevirt-api": "^0.0.35",
     "@patternfly/patternfly": "^4.171.1",
     "@patternfly/react-core": "^4.192.15",
     "classnames": "^2.3.1"
@@ -39,8 +39,8 @@
     "@babel/preset-env": "^7.16.8",
     "@babel/preset-react": "^7.16.7",
     "@babel/preset-typescript": "^7.16.7",
-    "@openshift-console/dynamic-plugin-sdk": "0.0",
-    "@openshift-console/dynamic-plugin-sdk-internal": "0.0",
+    "@openshift-console/dynamic-plugin-sdk": "0.0.5",
+    "@openshift-console/dynamic-plugin-sdk-internal": "^0.0.4",
     "@rollup/plugin-commonjs": "^21.0.1",
     "@rollup/plugin-node-resolve": "^13.1.3",
     "@rollup/plugin-typescript": "^8.3.0",
@@ -64,6 +64,7 @@
     "eslint-plugin-react-hooks": "^4.3.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "file-loader": "^6.2.0",
+    "http-proxy-middleware": "^2.0.3",
     "jest": "^27.4.7",
     "postcss": "^8.4.5",
     "prettier": "^2.5.1",
