+++ src/components/index.ts
@@ -1,1 +1,1 @@
-export * from './status';
+export * from './utils';
