+++ src/components/utils/README.md
@@ -1,0 +1,3 @@
+KubeVirt UI utility components used  to render general purpose components for KubeVirt UI elements for examples virtual machines status, time indicators.
+
+When needing a status component always look first at [PatternFly 4 guidelines](https://www.patternfly.org/v4/) and only if the component you are looking is missing, use KubeVirt UI components.
